requirements.txt
mcp-use
langchain-community
langchain-core
langchain-openai
fastapi 
browser-use
openai
langgraph
tavily-python
requests
python-dotenv
wikipedia
